Simplify new message modal
